Restructuring architecture final
