+++ src/main/java/com/hemebiotech/controller/getWriteSymptomsOccurrence.java
@@ -1,0 +1,4 @@
+package com.hemebiotech.controller;
+
+public class getWriteSymptomsOccurrence {
+}
