+++ src/main/java/com/hemebiotech/model/ISymptomReader.java
@@ -1,4 +1,4 @@
-package com.hemebiotech;
+package com.hemebiotech.model;
 
 import java.util.List;
